refactor(cloud): split client and organization modules
