+++ airbyte/cloud/workspaces.py
@@ -43,16 +43,12 @@ from typing import TYPE_CHECKING, Any, Literal, overload
 import yaml
 
 from airbyte import exceptions as exc
-from airbyte._util import api_imports, api_util, text_util
+from airbyte._util import api_util, text_util
 from airbyte._util.api_util import get_web_url_root
 from airbyte.cloud._credentials import (
-    CREDENTIALS_FILE_PATH,
-    CloudCredentials,
-    CloudLoginResult,
-    login_with_client_credentials,
+    _AirbyteCredentials,
     resolve_cloud_credentials,
 )
-from airbyte.cloud._credentials import logout as remove_credentials_file
 from airbyte.cloud.client_config import CloudClientConfig
 from airbyte.cloud.connections import CloudConnection
 from airbyte.cloud.connectors import (
@@ -60,408 +56,17 @@ from airbyte.cloud.connectors import (
     CloudSource,
     CustomCloudSourceDefinition,
 )
+from airbyte.cloud.organizations import CloudOrganization
 from airbyte.destinations.base import Destination
-from airbyte.exceptions import AirbyteError, AirbyteMissingResourceError
-from airbyte.secrets.base import SecretString
+from airbyte.exceptions import AirbyteError
 
 
 if TYPE_CHECKING:
     from collections.abc import Callable
 
     from airbyte._util.api_imports import JobResponse, WorkspaceResponse
+    from airbyte.secrets.base import SecretString
     from airbyte.sources.base import Source
-
-
-@dataclass(kw_only=True)
-class CloudClient:
-    """Authenticated client for Airbyte Cloud and self-managed Airbyte APIs."""
-
-    client_id: SecretString | None = None
-    client_secret: SecretString | None = None
-    bearer_token: SecretString | None = None
-    public_api_root: str = api_util.CLOUD_API_ROOT
-    config_api_root: str | None = None
-    organization_id: str | None = None
-
-    def __post_init__(self) -> None:
-        """Wrap provided secret values."""
-        if self.client_id is not None:
-            self.client_id = SecretString(self.client_id)
-        if self.client_secret is not None:
-            self.client_secret = SecretString(self.client_secret)
-        if self.bearer_token is not None:
-            self.bearer_token = SecretString(self.bearer_token)
-
-    @classmethod
-    def from_env(
-        cls,
-        *,
-        client_id: str | SecretString | None = None,
-        client_secret: str | SecretString | None = None,
-        bearer_token: str | SecretString | None = None,
-        organization_id: str | None = None,
-        public_api_root: str | None = None,
-        config_api_root: str | None = None,
-    ) -> CloudClient:
-        """Create a client from shared environment and credentials-file resolution."""
-        credentials = resolve_cloud_credentials(
-            client_id=client_id,
-            client_secret=client_secret,
-            bearer_token=bearer_token,
-            organization_id=organization_id,
-            public_api_root=public_api_root,
-            config_api_root=config_api_root,
-        )
-        return cls.from_credentials(credentials)
-
-    @classmethod
-    def from_auth(
-        cls,
-        *,
-        organization_id: str | None = None,
-        client_id: str | SecretString | None = None,
-        client_secret: str | SecretString | None = None,
-        bearer_token: str | SecretString | None = None,
-        public_api_root: str | None = None,
-        config_api_root: str | None = None,
-        credentials_file_path: Path = CREDENTIALS_FILE_PATH,
-    ) -> CloudClient:
-        """Create a client from explicit inputs, env vars, and credentials file."""
-        credentials = resolve_cloud_credentials(
-            organization_id=organization_id,
-            client_id=client_id,
-            client_secret=client_secret,
-            bearer_token=bearer_token,
-            public_api_root=public_api_root,
-            config_api_root=config_api_root,
-            credentials_file_path=credentials_file_path,
-        )
-        return cls.from_credentials(credentials)
-
-    def login(
-        self,
-        *,
-        interactive: bool | None = None,
-        credentials_file_path: Path = CREDENTIALS_FILE_PATH,
-    ) -> CloudLoginResult:
-        """Log in to Airbyte and persist local credentials."""
-        if interactive is True:
-            # TK-TODO: Implement and verify interactive browser login before merging this PR.
-            raise NotImplementedError("Interactive Airbyte Cloud login is not implemented.")
-        if self.client_id is not None and self.client_secret is not None:
-            return login_with_client_credentials(
-                client_id=str(self.client_id),
-                client_secret=str(self.client_secret),
-                airbyte_api_root=self.public_api_root,
-                config_api_root=self.config_api_root,
-                credentials_file_path=credentials_file_path,
-            )
-        if interactive is False:
-            raise exc.PyAirbyteInputError(
-                message="Client ID and client secret are both required.",
-                guidance="Provide both client ID and client secret for non-interactive login.",
-            )
-
-        # TK-TODO: Implement and verify interactive browser login before merging this PR.
-        raise NotImplementedError("Interactive Airbyte Cloud login is not implemented.")
-
-    def logout(
-        self,
-        *,
-        credentials_file_path: Path = CREDENTIALS_FILE_PATH,
-    ) -> None:
-        """Log out by removing locally stored credentials."""
-        remove_credentials_file(credentials_file_path=credentials_file_path)
-
-    @classmethod
-    def from_credentials(cls, credentials: CloudCredentials) -> CloudClient:
-        """Create a client from resolved Cloud credentials."""
-        return cls(
-            client_id=credentials.client_id,
-            client_secret=credentials.client_secret,
-            bearer_token=credentials.bearer_token,
-            public_api_root=credentials.public_api_root,
-            config_api_root=credentials.config_api_root,
-            organization_id=credentials.organization_id,
-        )
-
-    @classmethod
-    def from_explicit_credentials(
-        cls,
-        *,
-        client_id: str | SecretString | None = None,
-        client_secret: str | SecretString | None = None,
-        bearer_token: str | SecretString | None = None,
-        public_api_root: str | None = None,
-        config_api_root: str | None = None,
-        organization_id: str | None = None,
-    ) -> CloudClient:
-        """Create a client from explicit credential values only."""
-        return cls.from_credentials(
-            CloudCredentials(
-                client_id=SecretString(client_id) if client_id else None,
-                client_secret=SecretString(client_secret) if client_secret else None,
-                bearer_token=SecretString(bearer_token) if bearer_token else None,
-                public_api_root=public_api_root or api_util.CLOUD_API_ROOT,
-                config_api_root=config_api_root,
-                organization_id=organization_id,
-            )
-        )
-
-    def get_workspace(self, workspace_id: str | None = None) -> CloudWorkspace:
-        """Create a `CloudWorkspace` using this client's credentials."""
-        if not workspace_id:
-            raise exc.PyAirbyteInputError(
-                message="Workspace ID is required.",
-                guidance="Provide a workspace ID.",
-            )
-
-        return CloudWorkspace(
-            workspace_id=workspace_id,
-            client_id=self.client_id,
-            client_secret=self.client_secret,
-            bearer_token=self.bearer_token,
-            api_root=self.public_api_root,
-            config_api_root=self.config_api_root,
-        )
-
-    def list_workspaces(
-        self,
-        name: str | None = None,
-        *,
-        name_filter: Callable[[str], bool] | None = None,
-    ) -> list[api_imports.WorkspaceResponse]:
-        """List workspaces available to this client."""
-        return api_util.list_workspaces(
-            workspace_id="",
-            api_root=self.public_api_root,
-            name=name,
-            name_filter=name_filter,
-            client_id=self.client_id,
-            client_secret=self.client_secret,
-            bearer_token=self.bearer_token,
-        )
-
-    def list_workspaces_in_organization(
-        self,
-        organization_id: str | None = None,
-        *,
-        name_contains: str | None = None,
-        limit: int | None = None,
-    ) -> list[dict[str, object]]:
-        """List workspaces in an organization using the Config API."""
-        resolved_organization_id = organization_id or self.organization_id
-        if not resolved_organization_id:
-            raise exc.PyAirbyteInputError(
-                message="Organization ID is required.",
-                guidance="Provide an organization ID.",
-            )
-
-        return api_util.list_workspaces_in_organization(
-            organization_id=resolved_organization_id,
-            api_root=self.public_api_root,
-            config_api_root=self.config_api_root,
-            client_id=self.client_id,
-            client_secret=self.client_secret,
-            bearer_token=self.bearer_token,
-            name_contains=name_contains,
-            limit=limit,
-        )
-
-    def get_organization(
-        self,
-        organization_id: str | None = None,
-        *,
-        organization_name: str | None = None,
-    ) -> CloudOrganization:
-        """Resolve an organization by ID or exact name."""
-        if organization_id and organization_name:
-            raise exc.PyAirbyteInputError(
-                message="Provide either organization ID or organization name."
-            )
-        if not organization_id and not organization_name:
-            raise exc.PyAirbyteInputError(
-                message="Organization ID or organization name is required."
-            )
-
-        organizations = api_util.list_organizations_for_user(
-            api_root=self.public_api_root,
-            client_id=self.client_id,
-            client_secret=self.client_secret,
-            bearer_token=self.bearer_token,
-        )
-        if organization_id:
-            matching_organizations = [
-                organization
-                for organization in organizations
-                if organization.organization_id == organization_id
-            ]
-        else:
-            matching_organizations = [
-                organization
-                for organization in organizations
-                if organization.organization_name == organization_name
-            ]
-
-        if not matching_organizations:
-            raise AirbyteMissingResourceError(resource_type="organization")
-        if len(matching_organizations) > 1:
-            raise exc.PyAirbyteInputError(
-                message="Organization name matches multiple organizations."
-            )
-
-        organization = matching_organizations[0]
-
-        return CloudOrganization(
-            organization_id=organization.organization_id,
-            organization_name=organization.organization_name,
-            email=organization.email,
-            api_root=self.public_api_root,
-            config_api_root=self.config_api_root,
-            client_id=self.client_id,
-            client_secret=self.client_secret,
-            bearer_token=self.bearer_token,
-        )
-
-
-class CloudOrganization:
-    """Information about an organization in Airbyte Cloud.
-
-    This class provides lazy loading of organization attributes including billing status.
-    It is typically created via CloudWorkspace.get_organization().
-    """
-
-    def __init__(
-        self,
-        organization_id: str,
-        organization_name: str | None = None,
-        email: str | None = None,
-        *,
-        api_root: str = api_util.CLOUD_API_ROOT,
-        client_id: SecretString | None = None,
-        client_secret: SecretString | None = None,
-        bearer_token: SecretString | None = None,
-        config_api_root: str | None = None,
-    ) -> None:
-        """Initialize a CloudOrganization.
-
-        Args:
-            organization_id: The organization ID.
-            organization_name: Display name of the organization.
-            email: Email associated with the organization.
-            api_root: The API root URL.
-            client_id: OAuth client ID for authentication.
-            client_secret: OAuth client secret for authentication.
-            bearer_token: Bearer token for authentication (alternative to client credentials).
-            config_api_root: Optional Config API root URL.
-        """
-        self.organization_id = organization_id
-        """The organization ID."""
-
-        self._organization_name = organization_name
-        """Display name of the organization."""
-
-        self._email = email
-        """Email associated with the organization."""
-
-        self._api_root = api_root
-        self._config_api_root = config_api_root
-        self._client_id = client_id
-        self._client_secret = client_secret
-        self._bearer_token = bearer_token
-
-        # Cached organization info (billing, etc.)
-        self._organization_info: dict[str, Any] | None = None
-        # Flag to remember if fetching organization info failed (e.g., permission issues)
-        self._organization_info_fetch_failed: bool = False
-
-    def _fetch_organization_info(self, *, force_refresh: bool = False) -> dict[str, Any]:
-        """Fetch and cache organization info including billing status.
-
-        If fetching fails (e.g., due to permission issues), the failure is cached and
-        subsequent calls will return an empty dict without retrying.
-
-        Args:
-            force_refresh: If True, always fetch from the API even if cached.
-
-        Returns:
-            Dictionary containing organization info including billing data.
-            Returns empty dict if fetching failed or is not permitted.
-        """
-        # Reset failure flag if force_refresh is requested
-        if force_refresh:
-            self._organization_info_fetch_failed = False
-
-        # If we already know fetching failed, return empty dict without retrying
-        if self._organization_info_fetch_failed:
-            return {}
-
-        if not force_refresh and self._organization_info is not None:
-            return self._organization_info
-
-        try:
-            self._organization_info = api_util.get_organization_info(
-                organization_id=self.organization_id,
-                api_root=self._api_root,
-                config_api_root=self._config_api_root,
-                client_id=self._client_id,
-                client_secret=self._client_secret,
-                bearer_token=self._bearer_token,
-            )
-        except Exception:
-            # Cache the failure so we don't retry on subsequent property accesses
-            self._organization_info_fetch_failed = True
-            return {}
-        else:
-            return self._organization_info
-
-    @property
-    def organization_name(self) -> str | None:
-        """Display name of the organization."""
-        if self._organization_name is not None:
-            return self._organization_name
-        # Try to fetch from API if not set (returns empty dict on failure)
-        info = self._fetch_organization_info()
-        return info.get("organizationName")
-
-    @property
-    def email(self) -> str | None:
-        """Email associated with the organization."""
-        if self._email is not None:
-            return self._email
-        # Try to fetch from API if not set (returns empty dict on failure)
-        info = self._fetch_organization_info()
-        return info.get("email")
-
-    @property
-    def payment_status(self) -> str | None:
-        """Payment status of the organization.
-
-        Possible values: 'uninitialized', 'okay', 'grace_period', 'disabled', 'locked', 'manual'.
-        When 'disabled', syncs are blocked due to unpaid invoices.
-        Returns None if billing info is not available (e.g., due to permission issues).
-        """
-        info = self._fetch_organization_info()
-        return (info.get("billing") or {}).get("paymentStatus")
-
-    @property
-    def subscription_status(self) -> str | None:
-        """Subscription status of the organization.
-
-        Possible values: 'pre_subscription', 'subscribed', 'unsubscribed'.
-        Returns None if billing info is not available (e.g., due to permission issues).
-        """
-        info = self._fetch_organization_info()
-        return (info.get("billing") or {}).get("subscriptionStatus")
-
-    @property
-    def is_account_locked(self) -> bool:
-        """Whether the account is locked due to billing issues.
-
-        Returns True if payment_status is 'disabled'/'locked' or subscription_status is
-        'unsubscribed'. Defaults to False unless we have affirmative evidence of a locked state.
-        """
-        return api_util.is_account_locked(self.payment_status, self.subscription_status)
 
 
 @dataclass(init=False, kw_only=True)  # noqa: PLR0904  # Core cloud API facade.
@@ -501,8 +106,9 @@ class CloudWorkspace:
     """The Config API root URL."""
     bearer_token: SecretString | None
 
-    # Internal credentials object (set in __init__, excluded from repr)
-    _credentials: CloudClientConfig | None = field(default=None, init=False, repr=False)
+    # Internal credentials objects (set in __init__, excluded from repr)
+    _credentials: _AirbyteCredentials = field(init=False, repr=False)
+    _client_config: CloudClientConfig = field(init=False, repr=False)
 
     def __init__(
         self,
@@ -530,21 +136,33 @@ class CloudWorkspace:
                 guidance="Provide a workspace ID.",
             )
 
-        client = CloudClient.from_credentials(credentials)
+        self._credentials = credentials
         self.workspace_id = credentials.workspace_id or ""
-        self.client_id = client.client_id
-        self.client_secret = client.client_secret
-        self.bearer_token = client.bearer_token
-        self.api_root = client.public_api_root
-        self.config_api_root = client.config_api_root
+        self.client_id = credentials.client_id
+        self.client_secret = credentials.client_secret
+        self.bearer_token = credentials.bearer_token
+        self.api_root = credentials.public_api_root
+        self.config_api_root = credentials.config_api_root
 
         # Create internal CloudClientConfig object (validates mutual exclusivity)
-        self._credentials = CloudClientConfig(
+        self._client_config = CloudClientConfig(
             client_id=self.client_id,
             client_secret=self.client_secret,
             bearer_token=self.bearer_token,
             api_root=self.api_root,
             config_api_root=self.config_api_root,
+        )
+
+    @classmethod
+    def from_credentials(cls, credentials: _AirbyteCredentials) -> CloudWorkspace:
+        """Create a workspace from resolved credentials."""
+        return cls(
+            workspace_id=credentials.workspace_id,
+            client_id=credentials.client_id,
+            client_secret=credentials.client_secret,
+            bearer_token=credentials.bearer_token,
+            api_root=credentials.public_api_root,
+            config_api_root=credentials.config_api_root,
         )
 
     @classmethod
@@ -699,11 +317,7 @@ class CloudWorkspace:
         return CloudOrganization(
             organization_id=organization_id,
             organization_name=organization_name,
-            api_root=self.api_root,
-            config_api_root=self.config_api_root,
-            client_id=self.client_id,
-            client_secret=self.client_secret,
-            bearer_token=self.bearer_token,
+            credentials=self._credentials.with_organization_id(organization_id),
         )
 
     # Test connection and creds
@@ -1069,13 +683,15 @@ class CloudWorkspace:
         name_filter: Callable | None = None,
     ) -> list[WorkspaceResponse]:
         """List workspaces available to the current credentials."""
-        return CloudClient(
+        return api_util.list_workspaces(
+            workspace_id="",
+            api_root=self.api_root,
+            name=name,
+            name_filter=name_filter,
             client_id=self.client_id,
             client_secret=self.client_secret,
             bearer_token=self.bearer_token,
-            public_api_root=self.api_root,
-            config_api_root=self.config_api_root,
-        ).list_workspaces(name=name, name_filter=name_filter)
+        )
 
     def list_connections(
         self,
